Adding finch as my presentation layer
After playing around with a few libraries I've decided to go with finch
as a presentation layer for the budgeting system. The good thing is
adding a new presentation library is trivial (simply add a new module
for your presentation library and create your endpoints) so if you're
not wanting to use Finch then it is pretty easy to swap it out.

Libraries that I evaluated or have used and considered for this
application:
* http4s
* akka-http
* finch
* play