Adding persistence with doobie and in-memory database
Doobie integration for persistence and an in memory database testcase
for use with unit testing. The InMemoryDatabase allows for H2 database
with migrations (based on the flyway migrations used in the main
application)

The downside of this is it creates an interesting dependency chain in
the build and potential downsides is now there is 1 test database for
all unit tests.