Adding UserId foreign key
Apparently the database migrations were actually on gitignore? Adding a
UserId to the system to allow the easy transition to a system that can
be used for a single or many users. Right now i'm hard coding an initial
test user in the SQL file so that UserId(1) will exist. Later when I do
the UserRepository i'll remove this and adjust the test suite to create
test users