Converting CategoryService error to proper ADT
The Either[String, A] was a stop gap and i wanted to convert the String
to a proper error case. This was as good of a time as any.. which
highlighted my next thing that i need to do which is create a `find`
function on the Category Repository.

Converting from String highlighted this fact because now you have to
explicitly deal with all of your error states.