Adding MysqlError
Not the prettiest implementation but at least it DRYs the code a little
bit. The idea here is that i want to get all of the Mysql Errors behind
an ADT and then have a utility function that converts a SQLException
into a MysqlError.

The question i have in this implementation (among others) is should the
RepositoryError it corrisponds with be included into the MysqlError?
I've implemented it this way, but i may go back and revisit